(#7244) Add autosign_command for programmatic CSR autosigning
This commit adds support for delegating control of autosigning to an
outside command. This allows users to define their own criteria for
signing CSRs. This commit adds a new setting, 'autosign_command', for
the command to use for testing CSRs, and extends the logic of the CA to
test for autosigning based on both the 'autosign' setting and the
output of the autosign_command.

Original patch by Patrick Hemmer <patrick.hemmer@gmail.com>