(PP-724,#21922) Compare only whole secs when autoloading
Variations in nanosecs on a file's mtime were sufficient to cause the
autoloader to attempt to reload a file marked for autoloading prior to
this patch.  This seemed to manifest in two related issues PP-724 and
Redmine #21922, where (in some manner as yet unknown) a file's mtime
might end up differing only in nsecs.  This was causing issues when a
file, such as lib/puppet/util/instrumentation/listeners/log.rb, which is
marked for autoload even though it has constants which cannot be
reloaded, has nsecs changed by some unknown force, resulting in
failures on subsequent agent runs.