WIP some specs describing behavior with future parser
Not sure how useful this is.  Originally started with two sets of specs,
in future_compiler_spec and compiler_spec.  Tried consolidating into the
resource_expressions_spec because of all the duplication.  But the (a)
difficulty is that I'm trying to test different cases 'resource {
title_expression: }' and '$x = title_expression resource { $x: }', and,
at least in the current parser cases, error messages may differ
depending on whether it's a literal or a variable value.