(gh-101) Add local pre-suite steps
The acceptance pipeline was previously tied to use of beaker-puppet rake
tasks which would magically include a set of setup presuite steps from
that gem in order to install and prep ahead of running the tests/.

The beaker-puppet gem is tied to Perforce internals for obtaining
packages to test against. We're using Bolt in a prior step to install
agent/server packages as needed prior to running Beaker.

But we do need some of the setup steps to kick start certs/puppetserver
on the primary.

I'm extracting those steps from beaker-puppet into a local pre-suite
directory for our beaker runs.