(wip,gh-101) Add an acceptance workflow to gha
Uses the nested_vms action to prepare a cluster of one primary and one
agent, installs openvox-agent and openvox-server, then preps and
executes the acceptance/ suite with Beaker.

Currently this is using the action from
kvm_automation_tooling@generalize-dev-openvox-agent-acceptance
because it relies on newer kvm_automation_tooling functionality from
that branch that isn't in the v1 mainline used by nested_vms yet.