(gh-101) Add a pre-suite step to tune puppetserver memory use
The system package defaults to 2G.  Tuning it to half available system
memory to take advantage of more memory if available. Since there's no
puppetdb in these suite, there's no contention and Java can use half for
heap while still reserving half for off-heap and system usage. It could
be tuned higher depending on actual memory, but we don't need complex
heuristics here and won't be loading the puppetserver anyway. This is
just an attempt to give the server a little more breathing space to see
if it speeds up the tests at all.